refactor: reduce overly permissive access modifiers
