+++ Sources/TunnelMonitorKit/Mocking/TMTunnelProviderManager.swift
@@ -20,25 +20,38 @@ public protocol TMTunnelProviderManagerDelegate: AnyObject {
     func serviceStateChanged<ServiceInfo: Codable>(to state: ServiceInfo)
 }
 
-/// A base class allowing native and mock network extensions to be used interchangebly.
+/// A base class allowing native and mock network extensions to be used interchangebly. Its subclasses are responsible
+/// for creating, managing and monitoring a tunnel session. The host/container application should create an instance of
+/// `TMMockTunnelProviderManager` or `TMNativeTunnelProviderManager` using `TMMockTunnelProviderManagerFactory`, based
+/// on whether mocking is desired or not.
 public class TMTunnelProviderManager {
 
     init() {}
 
+    var session: TMTunnelProviderSession? { fatalError("Must override") }
+
+    var tunnelMonitor: TunnelMonitor { fatalError("Must override") }
+
+    /// Starts the tunnel.
     public func startTunnel() { }
+
+    /// Stops the tunnel.
     public func stopTunnel() { }
 
+    /// The delegate to which events are sent about the state of the tunnel provider and the service it provides.
     public weak var delegate: TMTunnelProviderManagerDelegate?
 
-    public var session: TMTunnelProviderSession? { fatalError("Must override") }
-
-    public var tunnelMonitor: TunnelMonitor { fatalError("Must override") }
-
     /// Current status of the network extension
-    public var tunnelStatus: NEVPNStatus { fatalError("Please use Mock or Native TunelProviderManager.") }
+    public var tunnelStatus: NEVPNStatus { fatalError("Must override") }
 
-    private var tunnelManager: NETunnelProviderManager { fatalError("please") }
-
+    /// Sends a message request to the tunnel provider.
+    ///
+    /// Automatically decoding any response to the type specified in the response handler. Use `send(_: Request)` when
+    /// you don't expect a response or don't care about decoding it or
+    /// its value.
+    /// - Parameters:
+    ///   - message: The message to send to the network extesion.
+    ///   - responseHandler: A response handler containing either a successfully decoded response, or an error.
     public func send<Request: Codable, Response: Codable>(
         message: Request,
         responseHandler: @escaping (Result<Response, TMCommunicationError>) -> Void
@@ -46,10 +59,23 @@ public class TMTunnelProviderManager {
         tunnelMonitor.send(message: message, responseHandler: responseHandler)
     }
 
+
+    /// Sends a message request to the tunnel provider, ignoring any response that is returned.
+    ///
+    /// Use `send(_: Request, _: @escaping (Result<Response, TMCommunicationError>) -> Void)` if you want to decode the
+    /// response to a specific Codable type.
+    /// - Parameters:
+    ///   - message: The message to send to the network extesion.
     public func send<Request: Codable>(message: Request) {
         tunnelMonitor.send(message: message) { (_: Result<Data, TMCommunicationError>) in }
     }
 
+    /// Starts monitoring the current network extension using state requests generated by the given request builder at
+    /// the specified interval.
+    /// - Parameters:
+    ///   - requestBuilder: The block responsible for constructing a status request.
+    ///   - responseHandler: The response handler which is invoked for every status response.
+    ///   - pollInterval: The interval at which to request status updates.
     public func startMonitoring<ServiceInfoRequest: Codable, ServiceInfoResponse: Codable>(
         withRequestBuilder requestBuilder: @escaping () -> ServiceInfoRequest,
         responseHandler: @escaping (Result<ServiceInfoResponse, TMCommunicationError>) -> Void,
@@ -67,12 +93,15 @@ public class TMTunnelProviderManager {
         )
     }
 
+    /// Stops monitoring the tunnel provider session.
     public func stopMonitoring() {
         tunnelMonitor.stopMonitoring()
     }
 
 }
 
+/// A mock tunnel provider manager which allows the functionality of a specific TMPacketTunnelProvider to be executed on
+/// host/container app targets and simulator target environments.
 public class TMMockTunnelProviderManager: TMTunnelProviderManager {
 
     private let provider: TMPacketTunnelProvider
@@ -89,7 +118,7 @@ public class TMMockTunnelProviderManager: TMTunnelProviderManager {
         }
     }
 
-    public init<UserConfiguration: Codable>(
+    init<UserConfiguration: Codable>(
         provider: TMPacketTunnelProvider,
         networkSettings: TMNetworkSettings,
         userConfiguration: UserConfiguration?
@@ -106,9 +135,9 @@ public class TMMockTunnelProviderManager: TMTunnelProviderManager {
         mockSession.setProvider(provider)
     }
 
+    override var session: TMTunnelProviderSession? { mockSession }
     override public var tunnelStatus: NEVPNStatus { currentTunnelStatus }
     override public var tunnelMonitor: TunnelMonitor { monitor }
-    override public var session: TMTunnelProviderSession? { mockSession }
 
     private func configureProvider(completionHandler: @escaping (TMTunnelConfigurationError?) -> Void) {
         provider.configureTunnel(
@@ -148,7 +177,8 @@ public class TMMockTunnelProviderManager: TMTunnelProviderManager {
     }
 }
 
-/// Class that manages the vpn network extension - invoked by the container app to launch the network extension
+/// Class that manages the vpn network extension. The native tunnel provider manager is essentially a wrapper around a
+/// real tunnel provider manager.
 public class TMNativeTunnelProviderManager: TMTunnelProviderManager {
 
     private let monitor = TunnelMonitor()
@@ -160,7 +190,7 @@ public class TMNativeTunnelProviderManager: TMTunnelProviderManager {
     override public var tunnelMonitor: TunnelMonitor { monitor }
 
     /// Initialise VPNManager, which also saves/updates a tunnel configuration to the phone's VPN settings
-    public init(provider: NETunnelProviderManager) {
+    init(provider: NETunnelProviderManager) {
         providerManager = provider
         super.init()
 
@@ -172,7 +202,7 @@ public class TMNativeTunnelProviderManager: TMTunnelProviderManager {
         )
     }
 
-    override public var session: TMTunnelProviderSession? {
+    override var session: TMTunnelProviderSession? {
         if let providerSession = providerManager.connection as? NETunnelProviderSession {
             return TMTunnelProviderSessionNative(nativeSession: providerSession)
         }
